Create PM-IRAS_v2_raw_code
A major updated version of the data process code for PM-IRAS. The code is created on Jupyter Lab so the format is ".ipynb".

To avoid any error due to the lack of a package, whenever use the code in this file, run the first three cells:
- Common package import
- Load data from a local directory
- Additional package for plot formatting and linear fit

The 4th Cell will help you automatically generate the repeated code text. You can save time loading multiple files or creating a data list without manually modifying the index of files. It doesn't interact with any data file. You can enter a code template (string) and print several lines of code as string. "File version" was developed earlier to generate codes from a local ".txt" file. "Interface version" is more recommended, because you can print the code you want at the original coding platform.

Three major data display functions have been updated. They allow you to plot single data, plot multiple data, and subtract two data. In the function "sglplot", in addition to the existing defined display range, you can also normalize the data, change the color of the curve, save the data as a local file, and draw vertical lines at specific wave numbers, which is mostly used in baseline correction. In the function "multiplot", in addition to the existing defined display range, you can also normalize the data, offset the data curve for clarity, change the color of the curve, save the data as a local file. In the function "subtr_plot", now you can scale the subtrahend.

New functions for baseline correction and preliminary peak fitting have been developed. A default cubic spines was used. You can change the order by defining a "k = n" in "splrep". The baseline corrected data name will add a suffix "_BL" before the last index. They are compatible with those data display functions mentioned above. The peak fitting function fits the peaks in a Gaussian shape. The "peak_fit2" records the fitting error as well as the peak area. The equations for error propagation are also provided. Those functions process a single data, print the details of fitting parameters, and display the fitted peaks in the same figure. If you want to compare the results from multiple data files, you'll need to write a for loop, add the fit details to a list, and then plot them separately. An example of the whole process was demonstrated.